fix: Indicator panel with TypedArrays
