+++ packages/studio-base/src/components/MessagePathSyntax/simpleGetMessagePathDataItems.test.ts
@@ -9,7 +9,7 @@ import { simpleGetMessagePathDataItems } from "./simpleGetMessagePathDataItems";
 
 describe("simpleGetMessagePathDataItems", () => {
   it("returns root message if topic matches", () => {
-    const message: MessageEvent = {
+    const message: MessageEvent<unknown> = {
       topic: "/foo",
       receiveTime: { sec: 0, nsec: 0 },
       sizeInBytes: 0,
@@ -20,8 +20,24 @@ describe("simpleGetMessagePathDataItems", () => {
     expect(simpleGetMessagePathDataItems(message, parseRosPath("/bar")!)).toEqual([]);
   });
 
+  it("supports TypedArray messages", () => {
+    const message: MessageEvent<unknown> = {
+      topic: "/foo",
+      receiveTime: { sec: 0, nsec: 0 },
+      sizeInBytes: 0,
+      schemaName: "datatype",
+      message: {
+        bar: new Uint32Array([3, 4, 5]),
+      },
+    };
+    expect(simpleGetMessagePathDataItems(message, parseRosPath("/foo.bar")!)).toEqual([
+      new Uint32Array([3, 4, 5]),
+    ]);
+    expect(simpleGetMessagePathDataItems(message, parseRosPath("/foo.bar[0]")!)).toEqual([3]);
+  });
+
   it("returns correct nested values", () => {
-    const message: MessageEvent = {
+    const message: MessageEvent<unknown> = {
       topic: "/foo",
       receiveTime: { sec: 0, nsec: 0 },
       sizeInBytes: 0,
@@ -55,7 +71,7 @@ describe("simpleGetMessagePathDataItems", () => {
   });
 
   it("returns nothing for missing fields", () => {
-    const message: MessageEvent = {
+    const message: MessageEvent<unknown> = {
       topic: "/foo",
       receiveTime: { sec: 0, nsec: 0 },
       sizeInBytes: 0,
@@ -66,7 +82,7 @@ describe("simpleGetMessagePathDataItems", () => {
   });
 
   it("throws for unsupported paths", () => {
-    const message: MessageEvent = {
+    const message: MessageEvent<unknown> = {
       topic: "/foo",
       receiveTime: { sec: 0, nsec: 0 },
       sizeInBytes: 0,
